+++ Changes.md
@@ -12,6 +12,7 @@ Features
 Improvements
 ------------
 
+- USD : Added automatic render-time translation of UsdPreviewSurface shaders to Arnold.
 - SceneReader :
   - Added support for Alembic visibility attributes.
   - Added loading of `treatAsPoint` and `treatAsLine` parameters from UsdLux lights.